Associated tests for CSimple grammar.
The tests are basically a duplicate of the C ones, but I've had to remove a lot
of the assert_nil tests, as the simple parser is very broad in what it allows.
Valid C will produce a parse tree, however invalid C may also produce a parse
tree; hence this parser should not be taken as a validation of any kind.